API/requirements.txt
django
mysql-connector-python
mysqlclient
djangorestframework

